+++ example_impl/example_services/example_service.go
@@ -12,5 +12,7 @@ import (
 type ExampleService interface {
 	services.Service
 
-	GetHelloWorldSocket() Socket
+	GetIpAddress() string
+
+	GetPort() int
 }